Restricted Pages
Reached the 2:49:00 timestamp.
Made it so that only the host could see the option to edit and delete their rooms. User's can also not re-login when already logged in, they will be redirected to the home page.

+++ studybud/base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -15,6 +17,9 @@ from .forms import RoomForm
 #]
 
 def loginPage(request):
+
+    if request.user.is_authenticated: #Will return the user to the home page if they try to login in when already logged in. 
+        return redirect('home')
 
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -61,6 +66,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -73,9 +79,13 @@ def createRoom(request):
     context = {'form' : form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room) 
+
+    if request.user != room.host: #User most be the host of the room in order to edit it 
+        return HttpResponse('You are not allowed here!!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room) #Updates the current room instance, and doesn't make a new post instead.
@@ -86,8 +96,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host: #User most be the host of the room in order to edit it 
+        return HttpResponse('You are not allowed here!!!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
